+++ lib/api_sketch/model.rb
@@ -155,6 +155,10 @@ module ApiSketch::Model
         self.all.find { |res| res.id == id }
       end
 
+      def find_by_http_method_and_path(http_method, path)
+        self.all.find { |res| res.http_method == http_method && res.path == path }
+      end
+
       def first
         self.all.first
       end
@@ -180,17 +184,30 @@ module ApiSketch::Model
         }
       end
 
+      def error_message(message)
+        # puts_error(message)
+        raise ::ApiSketch::Error, message
+      end
+
       def run_validations!
         unless self.action =~ /\A\w*\z/
-          message = "'#{self.action}' is invalid action value"
-          # puts_error(message)
-          raise ::ApiSketch::Error, message
+          error_message("'#{self.action}' is invalid action value")
         end
 
         if self.class.find(self.id)
-          message =  "'#{self.id}' is not unique id. Change values of 'namespace' and/or 'action' attributes"
-          # puts_error(message)
-          raise ::ApiSketch::Error, message
+          error_message("'#{self.id}' is not unique id. Change values of 'namespace' and/or 'action' attributes")
+        end
+
+        if self.http_method.nil? || self.http_method.empty?
+          error_message("request http_method can't be blank")
+        end
+
+        if self.path.nil? || self.path.empty?
+          error_message("request path can't be blank")
+        end
+
+        if self.class.find_by_http_method_and_path(self.http_method, self.path)
+          error_message("Route '#{self.http_method} #{self.path}' should be unique")
         end
       end