+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.1.0 / 2015-05-03
+# 0.1.1 / 2015-05-03
 
 - Bootstrap template: put all generated pages data into `{output_folder}/docs` directory. Left `assets` directory outside since it may clash with generated files/folders names.
 - Add endpoint search by conjuction of request `path` & `http_method` for API examples server (Works like normal api server now).